Demo15/requirements.txt
langchain
langchain-google-genai
google-generativeai
python-dotenv
